pyproject.toml
[project]
name = "bio-meta-ic"
version = "0.1.0"
description = "Biomorphic meta-learning integrated circuit simulator and trainer"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "numpy>=1.26",
    "scipy>=1.11",
    "networkx>=3.2",
    "matplotlib>=3.8",
    "pyyaml>=6.0",
    "tqdm>=4.66",
    "rich>=13.9",
    "cffi>=1.17",
    "graphviz>=0.20.3",
    "pygraphviz>=1.13"
]

[tool.black]
line-length = 100

[tool.isort]
profile = "black"
